refactor(console): remove old api-portal-header ajs component
