+++ gravitee-apim-console-webui/src/management/settings/api-portal-header/api-portal-header-edit-dialog/api-portal-header-edit-dialog.component.scss
