+++ gravitee-apim-console-webui/src/management/settings/api-portal-header/api-portal-header-edit-dialog/api-portal-header-edit-dialog.component.ts
@@ -17,7 +17,7 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 
-import { ApiPortalHeader } from '../../../../../entities/apiPortalHeader';
+import { ApiPortalHeader } from '../../../../entities/apiPortalHeader';
 
 export interface ApiPortalHeaderEditDialogResult {
   name: string;
