+++ gravitee-apim-console-webui/src/management/settings/api-portal-header/api-portal-header.harness..ts
