--- gravitee-apim-console-webui/src/management/settings/api-portal-header/new.api-portal-header.dialog.controller.ts
@@ -1,34 +1,0 @@
-/*
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { ApiPortalHeader } from '../../../entities/apiPortalHeader';
-import ApiHeaderService from '../../../services/apiHeader.service';
-
-function NewApiPortalHeaderDialogController(ApiHeaderService: ApiHeaderService, $mdDialog: angular.material.IDialogService) {
-  this.header = new ApiPortalHeader();
-
-  this.cancel = function () {
-    $mdDialog.cancel();
-  };
-
-  this.save = function () {
-    ApiHeaderService.create(this.header).then((response) => {
-      $mdDialog.hide(response.data);
-    });
-  };
-}
-NewApiPortalHeaderDialogController.$inject = ['ApiHeaderService', '$mdDialog'];
-
-export default NewApiPortalHeaderDialogController;
