--- gravitee-apim-console-webui/src/management/settings/api-portal-header/save.api-portal-header.dialog.html
@@ -1,38 +1,0 @@
-<!--
-
-    Copyright (C) 2015 The Gravitee team (http://gravitee.io)
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-            http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<md-dialog aria-label="{{$ctrl.header.id?'Update':'New'}} API Header">
-  <form name="form" ng-submit="$ctrl.save()" novalidate>
-    <md-dialog-content layout-padding>
-      <h4>{{$ctrl.header.id?'Update':'New'}} API Header</h4>
-      <div layout="column">
-        <md-input-container>
-          <label>Name</label>
-          <input ng-model="$ctrl.header.name" />
-        </md-input-container>
-        <md-input-container>
-          <label>Value</label>
-          <input ng-model="$ctrl.header.value" />
-        </md-input-container>
-      </div>
-    </md-dialog-content>
-    <md-dialog-actions layout="row">
-      <md-button ng-click="$ctrl.cancel()"> Cancel </md-button>
-      <md-button type="submit" class="md-raised md-primary" ng-disabled="form.$invalid"> {{$ctrl.header.id?'Update':'Create'}} </md-button>
-    </md-dialog-actions>
-  </form>
-</md-dialog>
